Implement new utility script wait-for-tcp.sh
This script utilizes netcat to wait until a tcp connection to the
specified host & port are available or until maximum number of attempts
have been attempted.

The script is now utilized in the docker-compose.yml for the services to
wait for rabbitmq to be available. This should make the startup more
consistent.